requirements.txt
# Core Python
python-dotenv
requests
GitPython

# GitHub API
PyGithub

# LLM & Vector Store (RAG)
torch
transformers
accelerate
bitsandbytes
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-text-splitters
python-dotenv
ollama

# Utilities
tqdm